fastapi/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "ollama",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "langchain-ollama",
    "langgraph",
    "llama-index",
    "torch",
    "transformers",
    "sentence-transformers",
    "opik",
    "loguru",
    'twikit',
    'linkedin-api',
    "pydantic",
    "crawl4ai",
    "python-dotenv",
    "qdrant-client",
    "numpy",
    "pymongo",
    "selenium",
    "chromedriver-autoinstaller",
    "brotli",
    "html2text",
    "beautifulsoup4>=4.14.3",
    "ddgs",
    "wikipedia>=1.4.0",
    "sqlmodel>=0.0.32",
    "duckduckgo-search>=8.1.1",
    "datasets>=4.5.0",
    "instructor",
    "soundfile",
    "grandalf",
    "chromadb>=1.4.1",
    "langsmith>=0.6.8",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.12",
]
